# Limit zlib output to 8 KB per iteration