--- a/src/decoder/zlib.rs
+++ b/src/decoder/zlib.rs
@@ -115,12 +115,12 @@ impl ZlibStream {
 
         let (buffer, filled) = image_data.borrow_mut();
         let output_limit = (filled + 8 * 1024).min(buffer.len());
-        let (in_consumed, out_consumed) =
-            self.state
-                .read(data, &mut buffer[..output_limit], filled, false)
-                .map_err(|err| {
-                    DecodingError::Format(FormatErrorInner::CorruptFlateStream { err }.into())
-                })?;
+        let (in_consumed, out_consumed) = self
+            .state
+            .read(data, &mut buffer[..output_limit], filled, false)
+            .map_err(|err| {
+                DecodingError::Format(FormatErrorInner::CorruptFlateStream { err }.into())
+            })?;
 
         self.started = true;
         let filled = filled + out_consumed;