--- a/src/decoder/zlib.rs
+++ b/src/decoder/zlib.rs
@@ -1,4 +1,4 @@
-use super::{stream::FormatErrorInner, DecodingError};
+use super::{stream::FormatErrorInner, unfiltering_buffer::UnfilteringBuffer, DecodingError};
 
 use fdeflate::Decompressor;
 
@@ -114,7 +114,7 @@ impl ZlibStream {
         }
 
         let (buffer, filled) = image_data.borrow_mut();
-        let output_limit = (filled + 8 * 1024).min(buffer.len());
+        let output_limit = (filled + UnfilteringBuffer::GROWTH_BYTES).min(buffer.len());
         let (in_consumed, out_consumed) = self
             .state
             .read(data, &mut buffer[..output_limit], filled, false)