@kornelski 2025-06-06
Why 8K? 

Is it more related to the deflate context window (an unlikely to change), or something else like cache size (therefore will need bumping up or down per platform)?

Can you use a constant for it so it's easier to spot and tune in the future?