@fintelia 2025-06-06
Added a constant to make it clear that the 8 KB is to match the amount `UnfilteringBuffer` grows by