@197g 2025-06-06
I still find this very odd. Why does reducing the amount of decoding we do at once improve performance? I'm going to be happy about the find nevertheless.