@Shnatsel 2025-06-06
I suspect it's because of the CPU cache. With this PR we decompress a chunk that fits into the cache and immediately filter it. By the time we're done with a 128kb chunk, the start of it no longer fits into L1 cache, so we have to spend time to load it from a slower cache again.

It wouldn't have mattered if we were simply outputting the zlib stream; but since we're immediately running further processing (filtering) over it, decompressing in chunks that fit into the fastest cache is beneficial.

This hypothesis can be verified using `perf` on Linux which provides counters for cache misses. I'd expect to see fewer L1 misses after this PR.